moved actor classes into Actors package and updated effected classes and updated module info
