+++ src/main/java/com/example/demo/Actors/ActiveActor.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.Actors;
 
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
