+++ src/main/java/com/example/demo/Actors/AdvancedEnemyPlane.java
@@ -1,4 +1,6 @@
-package com.example.demo;
+package com.example.demo.Actors;
+
+import com.example.demo.AudioPlayer;
 
 import java.util.ArrayList;
 import java.util.Collections;
