+++ src/main/java/com/example/demo/Actors/Boss.java
@@ -1,5 +1,6 @@
-package com.example.demo;
+package com.example.demo.Actors;
 
+import com.example.demo.AudioPlayer;
 import com.example.demo.LevelViews.LevelViewBoss;
 
 import java.util.ArrayList;
