+++ src/main/java/com/example/demo/Actors/BossProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.Actors;
 
 public class BossProjectile extends Projectile {
