+++ src/main/java/com/example/demo/Actors/Destructible.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.Actors;
 
 public interface Destructible {
