+++ src/main/java/com/example/demo/Actors/EnemyPlane.java
@@ -1,4 +1,6 @@
-package com.example.demo;
+package com.example.demo.Actors;
+
+import com.example.demo.AudioPlayer;
 
 public class EnemyPlane extends FighterPlane {
