+++ src/main/java/com/example/demo/Actors/EnemyProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.Actors;
 
 public class EnemyProjectile extends Projectile {
