+++ src/main/java/module-info.java
@@ -10,4 +10,5 @@ module com.example.demo {
     opens com.example.demo.ImageInitializers to javafx.fxml;
     opens com.example.demo.LevelViews to javafx.fxml;
     opens com.example.demo.controller to javafx.fxml;
+    opens com.example.demo.Actors to javafx.fxml;
 }
